+++ packages/hooks/src/orderly/useFundingRates.ts
@@ -24,6 +24,5 @@ export const useFundingRates = () => {
     }
   );
 
-  // @ts-ignore
   return createGetter<API.FundingRate>(data);
 };