+++ packages/hooks/CHANGELOG.md
@@ -1,16 +1,58 @@
 # @orderly.network/hooks
 
+## 1.1.9
+
+### Patch Changes
+
+- fix bug: app crashes when symbolInfo is null;
+- Updated dependencies
+  - @orderly.network/core@0.2.8
+  - @orderly.network/net@1.1.8
+  - @orderly.network/perp@2.0.8
+  - @orderly.network/types@0.2.8
+  - @orderly.network/utils@0.1.8
+
+## 1.1.8
+
+### Patch Changes
+
+- fix bug: window not defined;
+- Updated dependencies
+  - @orderly.network/core@0.2.7
+  - @orderly.network/net@1.1.7
+  - @orderly.network/perp@2.0.7
+  - @orderly.network/types@0.2.7
+  - @orderly.network/utils@0.1.7
+
+## 1.1.7
+
+### Patch Changes
+
+- fix bug: window is not defined
+
 ## 1.1.6
 
 ### Patch Changes
 
-- update version
-- Updated dependencies
+- 93a5ce6a: fix bugs;
+- Updated dependencies [93a5ce6a]
   - @orderly.network/types@0.2.6
   - @orderly.network/utils@0.1.6
   - @orderly.network/core@0.2.6
   - @orderly.network/perp@2.0.6
   - @orderly.network/net@1.1.6
+
+## 1.1.6-alpha.0
+
+### Patch Changes
+
+- fix bugs;
+- Updated dependencies
+  - @orderly.network/types@0.2.6-alpha.0
+  - @orderly.network/utils@0.1.6-alpha.0
+  - @orderly.network/core@0.2.6-alpha.0
+  - @orderly.network/perp@2.0.6-alpha.0
+  - @orderly.network/net@1.1.6-alpha.0
 
 ## 1.1.5