+++ packages/net/CHANGELOG.md
@@ -1,10 +1,28 @@
 # @orderly.network/net
 
+## 1.1.8
+
+### Patch Changes
+
+- fix bug: app crashes when symbolInfo is null;
+
+## 1.1.7
+
+### Patch Changes
+
+- fix bug: window not defined;
+
 ## 1.1.6
 
 ### Patch Changes
 
-- update version
+- 93a5ce6a: fix bugs;
+
+## 1.1.6-alpha.0
+
+### Patch Changes
+
+- fix bugs;
 
 ## 1.1.5