+++ packages/onboard/CHANGELOG.md
@@ -1,13 +1,47 @@
 # @orderly.network/onboard
 
+## 0.1.9
+
+### Patch Changes
+
+- fix bug: app crashes when symbolInfo is null;
+- Updated dependencies
+  - @orderly.network/hooks@1.1.9
+  - @orderly.network/types@0.2.8
+
+## 0.1.8
+
+### Patch Changes
+
+- fix bug: window not defined;
+- Updated dependencies
+  - @orderly.network/hooks@1.1.8
+  - @orderly.network/types@0.2.7
+
+## 0.1.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @orderly.network/hooks@1.1.7
+
 ## 0.1.6
 
 ### Patch Changes
 
-- update version
-- Updated dependencies
+- 93a5ce6a: fix bugs;
+- Updated dependencies [93a5ce6a]
   - @orderly.network/hooks@1.1.6
   - @orderly.network/types@0.2.6
+
+## 0.1.6-alpha.0
+
+### Patch Changes
+
+- fix bugs;
+- Updated dependencies
+  - @orderly.network/hooks@1.1.6-alpha.0
+  - @orderly.network/types@0.2.6-alpha.0
 
 ## 0.1.5